UnderworldTrial1
I removed some stuff from the file to make it fit, I'm pretty sure it was all non essential stuff that your own program should be able to fill in, but it is possible that removing some of the stuff may have buggered it up somehow. (Don't worry, I have the original safe and sound on my computer).